refactor: change filename BottomNav to BottomNavigation
