+++ src/pages/challenges/[id]/detail.tsx
@@ -4,7 +4,7 @@ import PageTitle from '@/components/common/PageTitle'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { challengesApi, challengesQueryKeys } from '@/api/challenges'
-import BottomNavigation from '@/components/common/BottomNav'
+import BottomNavigation from '@/components/common/BottomNavigation'
 import dayjs from 'dayjs'
 import { Button } from '@/components/ui/button'
 import { Dialog, DialogContent, DialogDescription } from '@/components/ui/dialog'
