+++ src/pages/posts/index.tsx
@@ -1,4 +1,4 @@
-import BottomNavigation from '@/components/common/BottomNav'
+import BottomNavigation from '@/components/common/BottomNavigation'
 import PostItem from '@/components/post-screen/post-item'
 import CategoryTab from '@/components/post-screen/category-tab'
 import { TabType } from '@/components/post-screen/category-tab/types'
